Rename test for clarity; test all thumbsnails
